removed unused logic
