+++ src/types.rs
@@ -10,6 +10,12 @@ pub enum ClientMessage {
     Ping(u64),  // Client sends timestamp
     Pong(u64),  // Server echoes timestamp
 }
+#[derive(Clone)]
+pub struct NetworkCondition {
+    pub latency_ms: i32,
+    pub packet_loss_percent: i32,
+    pub name: String,
+}
 
 #[derive(Serialize, Deserialize, Debug, Copy, Clone)]
 pub enum Direction {
@@ -30,6 +36,18 @@ pub struct PlayerInput {
 pub struct Position {
     pub x: i32,
     pub y: i32,
+}
+#[derive(Clone)]
+pub struct PositionSnapshot {
+    pub position: Position,
+    pub timestamp: u64,
+}
+
+#[derive(Debug, Clone)]
+pub struct InterpolatedPosition {
+    pub position: Position,
+    pub timestamp: f32,
+    pub sequence: u32,
 }
 
 #[derive(Serialize, Deserialize, Debug)]
